+++ cmd/main.go
@@ -236,6 +236,13 @@ func main() {
 		setupLog.Error(err, "unable to create controller", "controller", "HomeAssistantAutomation")
 		os.Exit(1)
 	}
+	if err := (&controller.HomeAssistantSceneReconciler{
+		Client: mgr.GetClient(),
+		Scheme: mgr.GetScheme(),
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "HomeAssistantScene")
+		os.Exit(1)
+	}
 	// +kubebuilder:scaffold:builder
 
 	if metricsCertWatcher != nil {